BlueGateFarm/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
PyMySQL==1.1.0
python-dotenv==1.0.0
Flask-Bootstrap==3.3.7.1
Flask-RESTful==0.3.10